refactor(debugging): migrate and consolidate debugging code
Moved debugging functions from Utils.cpp to a new file Debugging.cpp, creating a focused module for debugging utilities. Deleted Utils.cpp and its header, removing redundant code. Updated API_Version to 0.17.0 to reflect this change.

The refactoring enhances maintainability by consolidating all debugging-related functions into a dedicated module. This separation of concerns simplifies future updates and modifications related to debugging features.
