+++ Makefile
@@ -31,7 +31,7 @@ BUILD_TYPE	?= BIN
 LIB_MAJOR_VERSION	=	1
 LIB_MINOR_VERSION	=	1
 FW_Version	=	2.2.0
-API_Version	=	0.16.0
+API_Version	=	0.17.0
 # Flash size (KB).  Some low-end chips actually have more flash than advertised, use this to override.
 FLASH_SIZE	?=
 # Debugger optons, must be empty or GDB
@@ -279,8 +279,8 @@ DRONA_API =	API/Specifiers.cpp \
 						API-Src/FC-Data-Estimate.cpp \
 						API-Src/RC-Interface.cpp \
 						API-Src/Scheduler-Timer.cpp \
+						API-Src/Debugging.cpp \
 		    		API/XRanging.cpp \
-						API/Utils.cpp \
 						API-Src/Motor.cpp \
 						API/API-Utils.cpp \
 						API/RxConfig.cpp \
