+++ src/main/command/command.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include "../API/Utils.h"
+// #include "API-Src/Debugging.cpp"
 
 typedef enum {
